made independent of mainActivity
