+++ app/build.gradle
@@ -34,6 +34,6 @@ dependencies {
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.6.0'
 
     implementation "com.google.android.gms:play-services-base:17.1.0"
-    implementation "com.google.firebase:firebase-messaging:20.1.1"
+    implementation "com.google.firebase:firebase-messaging:20.1.2"
     implementation "com.google.android.gms:play-services-vision:19.0.0"
 }
