+++ app/src/main/java/com/example/gufyguber/MainActivity.java
@@ -1,26 +1,15 @@
 package com.example.gufyguber;
 
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.app.ActivityCompat;
-import androidx.core.content.ContextCompat;
 
-import android.Manifest;
 import android.content.Intent;
-import android.content.pm.PackageManager;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
-import android.widget.TextView;
-
-import com.google.android.gms.vision.barcode.Barcode;
 import com.google.firebase.firestore.FirebaseFirestore;
 
 public class MainActivity extends AppCompatActivity {
-    Button scan;
-    TextView result;
-
-    public static final int REQUEST_CODE = 100;
-    public static final int PERMISSION_REQUEST = 200;
+    Button start;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -29,35 +18,14 @@ public class MainActivity extends AppCompatActivity {
 
         FirebaseFirestore db = FirebaseFirestore.getInstance();
 
-        scan = findViewById(R.id.press);
-        result = findViewById(R.id.result);
+        start = findViewById(R.id.begin);
 
-        if(ContextCompat.checkSelfPermission(this, Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED) {
-            ActivityCompat.requestPermissions(this, new String [] {Manifest.permission.CAMERA}, PERMISSION_REQUEST);
-        }
-        scan.setOnClickListener(new View.OnClickListener() {
+        start.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent(MainActivity.this, Scan.class);
-                startActivityForResult(intent, REQUEST_CODE);
+                Intent start = new Intent(MainActivity.this, startScanQR.class);
+                startActivity(start);
             }
         });
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-
-        if(requestCode == REQUEST_CODE && resultCode == RESULT_OK) {
-            if(data != null) {
-                final Barcode barcode = data.getParcelableExtra("barcode");
-                result.post(new Runnable() {
-                    @Override
-                    public void run() {
-                        result.setText(barcode.displayValue);
-                    }
-                });
-            }
-        }
     }
 }
